+++ src/hal/actor/actor.py
@@ -9,8 +9,11 @@
 from __future__ import annotations
 
 import os
+from collections import deque
 
 from typing import TypeVar
+
+from sdssdb.peewee.sdss5db import targetdb
 
 from clu.legacy import LegacyActor
 
@@ -37,6 +40,44 @@ class HALActor(LegacyActor):
         self.version = __version__
 
         self.helpers = ActorHelpers(self)
+
+        # A double ended queue to store the last two fields. Elements are tuples
+        # with the format (field_id, is_cloned, is_rm_or_aqmes)
+        self.field_queue = deque(maxlen=2)
+        self.models["jaeger"]["configuration_loaded"].register_callback(
+            self._configuration_loaded
+        )
+
+    def _configuration_loaded(self, key):
+        """Processes a new configuration."""
+
+        design_id = key.value[1]
+        field_id = key.value[2]
+        is_cloned = key.value[9]
+
+        if design_id is None or design_id < 0 or field_id is None or field_id < 0:
+            self.field_queue.append((field_id, is_cloned, False))
+            return
+
+        if targetdb.database.connected is False:
+            self.write("w", {"error": "Cannot connect to database."})
+        else:
+            design_mode_label = (
+                targetdb.Design.select(targetdb.Design.design_mode)
+                .where(targetdb.Design.design_id == design_id)
+                .scalar()
+            )
+            if design_mode_label is None:
+                self.write(
+                    "w",
+                    {"error": f"Cannot find design_mode_label for design {design_id}"},
+                )
+            else:
+                if design_mode_label in ["dark_monit", "dark_rm"]:
+                    self.field_queue.append((field_id, is_cloned, True))
+                    return
+
+        self.field_queue.append((field_id, is_cloned, False))
 
 
 class ActorHelpers: